refactor(domain): enum 型参照を Prisma 生成物から domain/types へ統一
@/generated/prisma の enum 型 (Role / TicketStatus / Priority /
HistoryField) を直接 import していた非アダプタ層 11 ファイルを、正準
(SSOT) である @/domain/types 参照に統一する。すべて型のみ (import type)
依存のため実行時挙動は不変。

CLAUDE.md の「Prisma を直接 import するのはアダプタ層のみ / ドメインは
フレームワーク非依存」方針に沿った整理リファクタ。
