+++ src/features/tickets/components/PrioritySelect.tsx
@@ -7,7 +7,7 @@ import { updateTicketPriority } from '@/features/tickets/actions/update-ticket';
 // 優先度の日本語ラベル
 import { PRIORITY_LABELS } from '@/lib/constants';
 // 優先度の型 (Prisma 生成)
-import type { Priority } from '@/generated/prisma';
+import type { Priority } from '@/domain/types';
 
 // プルダウンに並べる優先度の順序
 const ALL_PRIORITIES: Priority[] = ['Low', 'Medium', 'High'];
